Fix restore from an old backup version
trimEntriesForExpiredMessages() has to be called after potentially upgrading the
database restored from a backup save with onUpgrade() since it tries to update
it with requests based on the newest schemas.
